Fix ultrasound pulse measurement overflowing at long distance This occurs above ~5.6m

+++ resources/kit/arduino-fw.ino
@@ -61,7 +61,7 @@ void command_ultrasound() {
   digitalWrite(pulse, LOW);
 
   // measure the echo time on the echo pin
-  int duration = pulseIn(echo, HIGH, 60000);
+  long duration = pulseIn(echo, HIGH, 60000);
   Serial.print(microsecondsToMm(duration));
 }
